1.8.0-20241111

+++ client/src/java/CSetting.java
@@ -35,8 +35,8 @@ public class CSetting {
     public native static String getString(String name, String defaultvalue);
     public native static void   setString(String name, String value);
 
-    public native static int    getNumber(String name, int defaultvalue);
-    public native static void   setNumber(String name, int value);
+    public native static long   getNumber(String name, long defaultvalue);
+    public native static void   setNumber(String name, long value);
 
     public native static void   save();
 }
